update and fix loi

+++ 3.PresentationLayers/Views/Form_Sach.cs
@@ -43,6 +43,7 @@ namespace _3.PresentationLayers.Views
             {
                 dtg_Show.Rows.Add(stt++, x.Id, x.Ma, x.Ten, x.TrangThai == 0 ? "Không hoạt động" : "Hoạt động");
             }
+            dtg_Show.AllowUserToAddRows = false;
         }
         private void Load()
         {
@@ -94,6 +95,7 @@ namespace _3.PresentationLayers.Views
         /*-------------------------------------------------------------------------*/
         private void dtg_Show_CellClick(object sender, DataGridViewCellEventArgs e)
         {
+            if (e.RowIndex == -1) return;
             SelectID = Guid.Parse(dtg_Show.CurrentRow.Cells[1].Value.ToString());
             tb_ma.Text = dtg_Show.CurrentRow.Cells[2].Value.ToString();
             tb_ten.Text = dtg_Show.CurrentRow.Cells[3].Value.ToString();
@@ -101,20 +103,47 @@ namespace _3.PresentationLayers.Views
         }
         private void btn_Them_Click(object sender, EventArgs e)
         {
-            MessageBox.Show(_iSachService.Add(GetDataFromGui_Them()));
-            ResetForm();
+            DialogResult dialogResult = MessageBox.Show("Bạn có chắc muốn thêm đối tượng không?", "Thông báo", MessageBoxButtons.YesNo);
+            if (dialogResult == DialogResult.Yes)
+            {
+                if (tb_ma.Text == "" || tb_ten.Text == "" || cbb_trangthai.Text == "--Chọn--")
+                {
+                    MessageBox.Show("Bạn chưa nhập đầy đủ thông tin", "Thông báo", MessageBoxButtons.OK);
+                    return;
+                }
+                MessageBox.Show(_iSachService.Add(GetDataFromGui_Them()));
+                ResetForm();
+            }
+            else return;
         }
 
         private void btn_Sua_Click(object sender, EventArgs e)
         {
-            MessageBox.Show(_iSachService.Update(GetDataFromGui_Sua_Xoa()));
-            ResetForm();
+            DialogResult dialogResult = MessageBox.Show("Bạn có chắc muốn cập nhật đối tượng không?", "Thông báo", MessageBoxButtons.YesNo);
+            if (dialogResult == DialogResult.Yes)
+            {
+                if (tb_ma.Text == "" || tb_ten.Text == "" || cbb_trangthai.Text == "--Chọn--")
+                {
+                    MessageBox.Show("Bạn chưa nhập đầy đủ thông tin", "Thông báo", MessageBoxButtons.OK);
+                    return;
+                }
+                MessageBox.Show(_iSachService.Update(GetDataFromGui_Sua_Xoa()));
+                ResetForm();
+            }
+            else return;
+
         }
 
         private void btn_Xoa_Click(object sender, EventArgs e)
         {
-            MessageBox.Show(_iSachService.Delete(GetDataFromGui_Sua_Xoa()));
-            ResetForm();
+            DialogResult dialogResult = MessageBox.Show("Bạn có chắc muốn xóa đối tượng không?", "Thông báo", MessageBoxButtons.YesNo);
+            if (dialogResult == DialogResult.Yes)
+            {
+                MessageBox.Show(_iSachService.Delete(GetDataFromGui_Sua_Xoa()));
+                ResetForm();
+            }
+            else return;
+
         }
 
         private void btn_Reset_Click(object sender, EventArgs e)
